+++ packages/core/types/index.d.ts
@@ -2,6 +2,7 @@ import Breadcrumb from "./breadcrumb";
 import Client from "./client";
 import Event from "./event";
 import Session from "./session";
+import BugsnagStatic from "./bugsnag";
 
 export * from "./common";
-export { Breadcrumb, Client, Event, Session };
+export { Breadcrumb, Client, Event, Session, BugsnagStatic };