+++ test/browser/features/fixtures/plugin_angular/ng/src/app/bugsnag.ts
@@ -1,9 +1,9 @@
-import bugsnag from '@bugsnag/browser';
+import Bugsnag from '@bugsnag/js';
 
 var ENDPOINT = decodeURIComponent(window.location.search.match(/ENDPOINT=([^&]+)/)[1])
 var API_KEY = decodeURIComponent(window.location.search.match(/API_KEY=([^&]+)/)[1])
 
-const bugsnagClient = bugsnag({
+const bugsnagClient = Bugsnag.createClient({
   apiKey: API_KEY,
   endpoints: {
     notify: ENDPOINT,