+++ test/expo/features/fixtures/test-app/app/console_breadcrumbs.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { View, Button } from 'react-native'
 import { endpoints } from './bugsnag'
-import bugsnag from '@bugsnag/expo'
+import Bugsnag from '@bugsnag/expo'
 
 export default class ConsoleBreadcrumbs extends Component {
   constructor(props) {
@@ -14,7 +14,7 @@ export default class ConsoleBreadcrumbs extends Component {
 
   defaultConsoleBreadcrumbsBehaviour = () => {
     this.triggerConsoleBreadcrumbsError(
-      bugsnag({
+      Bugsnag.createClient({
         endpoints: endpoints,
         autoDetectErrors: false,
         autoTrackSessions: false
@@ -25,7 +25,7 @@ export default class ConsoleBreadcrumbs extends Component {
 
   disabledConsoleBreadcrumbsBehaviour = () => {
     this.triggerConsoleBreadcrumbsError(
-      bugsnag({
+      Bugsnag.createClient({
         endpoints: endpoints,
         autoDetectErrors: false,
         autoTrackSessions: false,
@@ -37,7 +37,7 @@ export default class ConsoleBreadcrumbs extends Component {
 
   disabledAllConsoleBreadcrumbsBehaviour = () => {
     this.triggerConsoleBreadcrumbsError(
-      bugsnag({
+      Bugsnag.createClient({
         endpoints: endpoints,
         autoDetectErrors: false,
         autoTrackSessions: false,
@@ -49,7 +49,7 @@ export default class ConsoleBreadcrumbs extends Component {
 
   overrideConsoleBreadcrumbsBehaviour = () => {
     this.triggerConsoleBreadcrumbsError(
-      bugsnag({
+      Bugsnag.createClient({
         endpoints: endpoints,
         autoDetectErrors: false,
         autoTrackSessions: false,