+++ test/node/features/fixtures/project_root/scenarios/project-root-default.js
@@ -1,7 +1,7 @@
-var bugsnag = require('@bugsnag/node')
+var Bugsnag = require('@bugsnag/node')
 var lodash = require('lodash')
 
-var bugsnagClient = bugsnag({
+Bugsnag.init({
   apiKey: process.env.BUGSNAG_API_KEY,
   endpoints: {
     notify: process.env.BUGSNAG_NOTIFY_ENDPOINT,
@@ -12,5 +12,5 @@ var bugsnagClient = bugsnag({
 // the purpose of this seemingly pointless throttle call is just to make sure the
 // error has a stackframe from inside node_modules
 lodash.throttle(function () {
-  bugsnagClient.notify(new Error('project root'))
+  Bugsnag.notify(new Error('project root'))
 })()