+++ test/node/features/fixtures/sessions/scenarios/start-session-100.js
@@ -1,5 +1,5 @@
-var bugsnag = require('@bugsnag/node')
-var bugsnagClient = bugsnag({
+var Bugsnag = require('@bugsnag/node')
+Bugsnag.init({
   apiKey: process.env.BUGSNAG_API_KEY,
   endpoints: {
     notify: process.env.BUGSNAG_NOTIFY_ENDPOINT,
@@ -8,7 +8,7 @@ var bugsnagClient = bugsnag({
   sessionSummaryInterval: 1000
 })
 
-for (var i = 0; i < 100; i++) bugsnagClient.startSession()
+for (var i = 0; i < 100; i++) Bugsnag.startSession()
 
 // keep the process open a little bit so that the session has time to get sent
 setTimeout(function () {}, 5000)