+++ test/node/features/fixtures/surrounding_code/scenarios/send-code-default.js
@@ -1,5 +1,5 @@
-var bugsnag = require('@bugsnag/node')
-var bugsnagClient = bugsnag({
+var Bugsnag = require('@bugsnag/node')
+Bugsnag.init({
   apiKey: process.env.BUGSNAG_API_KEY,
   endpoints: {
     notify: process.env.BUGSNAG_NOTIFY_ENDPOINT,
@@ -17,7 +17,7 @@ function add (a) {
 add(5)(2) // -> 7
 subtract(5)(2) // -> 3
 
-bugsnagClient.notify(new Error('surround me'))
+Bugsnag.notify(new Error('surround me'))
 
 function subtract (a) {
   return function (b) {