+++ test/node/features/fixtures/unhandled/scenarios/contextualize.js
@@ -1,6 +1,6 @@
 var fs = require('fs')
-var bugsnag = require('@bugsnag/node')
-var bugsnagClient = bugsnag({
+var Bugsnag = require('@bugsnag/node')
+Bugsnag.init({
   apiKey: process.env.BUGSNAG_API_KEY,
   endpoints: {
     notify: process.env.BUGSNAG_NOTIFY_ENDPOINT,
@@ -8,7 +8,7 @@ var bugsnagClient = bugsnag({
   }
 })
 
-var contextualize = bugsnagClient.getPlugin('contextualize')
+var contextualize = Bugsnag.getPlugin('contextualize')
 contextualize(function () {
   fs.createReadStream('does not exist')
 }, function (event) {